Find the MOV opcode dynamically
The code had previously hard coded the MOV opcode to be #299. We
change that to find the MOV opcode dynamically. This change makes
the code more flexible, but will probably break the analysis code
which assumes a fixed relation of opcodes to opcode numbers (since
it is represented as a Haskell enumeration).