Change JumpMap to keep pointers to JumpEntry
The previous definition of the map used the actual JumpEntry
objects as the value type. I think this was causing a lot of
slowness becuase the JumpEntries were being copied all over
the place due to the semantics of map lookup. These copies
were expensive since they contain a std::set.