Fix two bug in ibmix tool
Bug 1. The case statement for counting branches was missing
the break statement so we were overcounting. For each call
we would also count that as a jump and a return. For each
jump we would also count that as a return.

Bug 2. The histogram was not cleared at the top of the loop.
This caused us to count the static returns as dynamic jumps.